+++ lib/src/widgets/nes_icon16.dart
@@ -1,0 +1,8 @@
+// ignore_for_file: lines_longer_than_80_chars
+import 'package:nes_ui/nes_ui.dart';
+
+/// Built in library of 16x16 icons for the Flutter Nes Design library.
+/// Use [NesIcons] to access the icons.
+class NesIcons16 {
+  NesIcons16._();
+}