+++ lib/src/widgets/widgets.dart
@@ -10,6 +10,7 @@ export 'nes_dropshadow.dart';
 export 'nes_file_explorer.dart';
 export 'nes_fixed_viewport.dart';
 export 'nes_icon.dart';
+export 'nes_icon16.dart';
 export 'nes_icon_badge.dart';
 export 'nes_icon_button.dart';
 export 'nes_input.dart';