+++ example/lib/gallery/sections/sections.dart
@@ -15,6 +15,7 @@ export 'key_icons.dart';
 export 'loading_indicators.dart';
 export 'running_text.dart';
 export 'screen_transitions.dart';
+export 'section_header.dart';
 export 'selection_list.dart';
 export 'single_child_scroll_view.dart';
 export 'snackbars.dart';